2x longer TICKs or your money back, guaranteed

+++ include/clock.h
@@ -1,6 +1,6 @@
 #pragma once
 
-typedef unsigned long TICK;
+typedef unsigned long long TICK;
 
 struct CLOCK
 {
